Migrate to uv workspaces for cluster-manager and fix some issues from previous types added
