+++ cluster_orchestrator/cluster-manager/Dockerfile
@@ -1,46 +1,54 @@
-FROM ghcr.io/astral-sh/uv:python3.10-trixie
-LABEL org.opencontainers.image.source https://github.com/oakestra/oakestra
+ARG BASE_DIR=/oakestra
+ARG APP_DIR=$BASE_DIR/cluster_orchestrator/cluster-manager
+ARG LIB_DIR=$BASE_DIR/libraries
 
-RUN apt-get update --yes \
-    && apt-get install --yes --no-install-recommends \
-      curl \
-      build-essential \
-      git
+FROM ghcr.io/astral-sh/uv:python3.10-trixie-slim AS builder
 
-COPY . /app
-WORKDIR /app
+ARG BASE_DIR
+ARG APP_DIR
+ARG LIB_DIR
 
-ENV UV_NO_DEV=1
-RUN uv sync --locked
+ENV UV_COMPILE_BYTECODE=1 \
+    UV_LINK_MODE=copy \
+    UV_NO_DEV=1
 
-# Install appropriate requirements
-ARG LIB_BRANCH=develop
-ENV LIB_BRANCH=${LIB_BRANCH}
+COPY pyproject.toml uv.lock $BASE_DIR
 
-RUN uv run -m grpc_tools.protoc -I. --python_out=. --grpc_python_out=. --pyi_out=. proto/clusterRegistration.proto
+WORKDIR $BASE_DIR
+RUN uv sync --frozen --no-install-workspace --package cluster-manager
 
-# ENV FLASK_APP=scheduler.py
-ENV FLASK_ENV=development
-# TRUE for verbose logging
-ENV FLASK_DEBUG=FALSE
+COPY libraries/ $LIB_DIR/
+COPY cluster_orchestrator/cluster-manager/ $APP_DIR/
 
-ENV SYSTEM_MANAGER_URL=3.120.37.66
-ENV SYSTEM_MANAGER_PORT=10000
-ENV SYSTEM_MANAGER_GRPC_PORT=50052
+WORKDIR $APP_DIR
+RUN uv run --no-sync -m grpc_tools.protoc \
+        -Isrc \
+        --python_out=src \
+        --grpc_python_out=src \
+        --pyi_out=src \
+        src/cluster_manager/proto/cluster_registration.proto \
+    && uv sync --locked --no-editable
 
-ENV CLUSTER_SERVICE_MANAGER_URL=localhost
-ENV CLUSTER_SERVICE_MANAGER_PORT=10010
+FROM python:3.10-slim-trixie
 
-ENV CLUSTER_MONGO_URL=localhost
-ENV CLUSTER_MONGO_PORT=10007
+ARG BASE_DIR
 
-ENV CLUSTER_SCHEDULER_URL=localhost
-ENV CLUSTER_SCHEDULER_PORT=5555
+COPY --from=builder $BASE_DIR/.venv $BASE_DIR/.venv
 
-ENV MQTT_BROKER_URL=localhost
-ENV MQTT_BROKER_PORT=10003
-ENV MY_PORT=10100
+ENV PATH="$BASE_DIR/.venv/bin:$PATH" \
+    FLASK_DEBUG=0 \
+    SYSTEM_MANAGER_URL=3.120.37.66 \
+    SYSTEM_MANAGER_PORT=10000 \
+    SYSTEM_MANAGER_GRPC_PORT=50052 \
+    CLUSTER_SERVICE_MANAGER_ADDR=localhost \
+    CLUSTER_SERVICE_MANAGER_PORT=10010 \
+    CLUSTER_MONGO_URL=localhost \
+    CLUSTER_MONGO_PORT=10007 \
+    CLUSTER_SCHEDULER_URL=localhost \
+    CLUSTER_SCHEDULER_PORT=5555 \
+    MQTT_BROKER_URL=localhost \
+    MQTT_BROKER_PORT=10003 \
+    MY_PORT=10100 \
+    LOG_LEVEL=INFO
 
-ENV LOG_LEVEL=INFO
-
-CMD gunicorn --bind [::]:${MY_PORT} cluster_manager:app
+CMD ["gunicorn", "--bind", "[::]:10100", "cluster_manager:app"]
