--- cluster_orchestrator/cluster-manager/blueprints/__init__.py
@@ -1,5 +1,0 @@
-from blueprints.cluster_blueprints import clusterblp
-from blueprints.service_blueprints import schedulingblp, serviceblp
-from blueprints.worker_blueprints import workerblp
-
-blueprints = [serviceblp, schedulingblp, workerblp, clusterblp]
