+++ cluster_orchestrator/cluster-manager/src/cluster_manager/__init__.py
@@ -1,0 +1,3 @@
+from .main import app
+
+__all__ = ["app"]
