+++ cluster_orchestrator/cluster-manager/src/cluster_manager/blueprints/__init__.py
@@ -1,0 +1,5 @@
+from .cluster_blueprints import clusterblp
+from .service_blueprints import schedulingblp, serviceblp
+from .worker_blueprints import workerblp
+
+blueprints = [serviceblp, schedulingblp, workerblp, clusterblp]
