+++ cluster_orchestrator/cluster-manager/src/cluster_manager/blueprints/worker_blueprints.py
@@ -1,0 +1,48 @@
+import logging
+
+from bson import json_util
+from flask import Response, request
+from flask.views import MethodView
+from flask_smorest import Blueprint, abort
+
+from resource_abstractor_client import candidate_operations
+from ..app_config import CONFIG
+
+logger = logging.getLogger("cluster_manager")
+
+# ........ Functions for job management ...............#
+# ......................................................#
+
+workerblp = Blueprint(
+    "Service operations",
+    "service",
+    url_prefix="/api/node",
+    description="Node registration operations",
+)
+
+
+@workerblp.route("/register")
+class ServiceController(MethodView):
+    @workerblp.response(
+        200,
+        {},
+        content_type="application/json",
+    )
+    def post(self):
+        logger.debug("Incoming Request /api/node/register - to register node")
+        data = request.json  # get POST body
+        data.get("token")  # registration_token
+        # TODO(GB): check and generate tokens
+        data["candidate_name"] = data.get("host", "")
+        logger.debug("Candidate Name: " + data.get("candidate_name"))
+        worker = candidate_operations.create_candidate(data)
+        if worker is None:
+            logger.error("Failed to register node")
+            abort(500, RuntimeError("Failed to register node"))
+
+        worker_id = str(worker["_id"])
+        response = {
+            "id": str(worker_id),
+            "MQTT_BROKER_PORT": CONFIG.mqtt_broker_port,
+        }
+        return Response(json_util.dumps(response), mimetype="application/json")
