+++ cluster_orchestrator/cluster-manager/src/cluster_manager/models/worker.py
