+++ cluster_orchestrator/cluster-manager/src/cluster_manager/proto/cluster_registration.proto
