+++ libraries/oakestra-utils/pyproject.toml
@@ -1,0 +1,3 @@
+[project]
+name = "oakestra-utils"
+version = "0.0.0"
