--- libraries/oakestra_utils_library/.python-version
@@ -1,1 +1,0 @@
-3.10.13
