--- libraries/oakestra_utils_library/pyproject.toml
@@ -1,13 +1,0 @@
-[tool.poetry]
-name = "oakestra_utils"
-version = "0.2.1"
-description = "A library containing common utilities for Oakestra"
-authors = ["Alexander Malyuk <malyuk.alexander1999@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
