+++ libraries/resource-abstractor-client/src/resource_abstractor_client/client_helper.py
