--- libraries/resource_abstractor_client/MANIFEST.in
@@ -1,2 +1,0 @@
-graft resource_abstractor_client
-global-include *.py
